CIF-1181: Refactor dataLayer clases  - Refactor ID generation for product related components
