+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/internal/datalayer/ComponentDataImpl.java
@@ -11,7 +11,7 @@
  *    governing permissions and limitations under the License.
  *
  ******************************************************************************/
-package com.adobe.cq.commerce.core.components.internal.models.v1.datalayer;
+package com.adobe.cq.commerce.core.components.internal.datalayer;
 
 import java.util.Calendar;
 import java.util.Date;
