+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/internal/datalayer/DataLayerComponent.java
@@ -11,7 +11,7 @@
  *    governing permissions and limitations under the License.
  *
  ******************************************************************************/
-package com.adobe.cq.commerce.core.components.internal.models.v1.datalayer;
+package com.adobe.cq.commerce.core.components.internal.datalayer;
 
 import javax.inject.Inject;
 
@@ -21,7 +21,7 @@ import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.caconfig.ConfigurationBuilder;
 
-import com.adobe.cq.commerce.core.components.models.datalayer.CategoryData;
+import com.adobe.cq.commerce.core.components.datalayer.CategoryData;
 import com.adobe.cq.wcm.core.components.models.datalayer.ComponentData;
 import com.fasterxml.jackson.annotation.JsonIgnore;
