+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/internal/models/v1/product/ProductRetriever.java
@@ -57,7 +57,6 @@ class ProductRetriever extends AbstractProductRetriever {
     private SimpleProductQueryDefinition generateSimpleProductQuery() {
         return q -> {
             q.sku()
-                .id()
                 .name()
                 .description(d -> d.html())
                 .image(i -> i.label().url())
@@ -84,7 +83,6 @@ class ProductRetriever extends AbstractProductRetriever {
     protected ProductInterfaceQueryDefinition generateProductQuery() {
         return q -> {
             q.sku()
-                .id()
                 .name()
                 .description(d -> d.html())
                 .image(i -> i.label().url())
