+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/internal/models/v1/searchbar/SearchbarImpl.java
@@ -19,7 +19,7 @@ import javax.inject.Inject;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.models.annotations.Model;
 
-import com.adobe.cq.commerce.core.components.internal.models.v1.datalayer.DataLayerComponent;
+import com.adobe.cq.commerce.core.components.internal.datalayer.DataLayerComponent;
 import com.adobe.cq.commerce.core.components.models.searchbar.Searchbar;
 import com.adobe.cq.commerce.core.components.utils.SiteNavigation;
 import com.day.cq.wcm.api.Page;
