+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/productcarousel/package-info.java
@@ -12,7 +12,7 @@
  *
  ******************************************************************************/
 
-@Version("2.0.0")
+@Version("2.1.0")
 package com.adobe.cq.commerce.core.components.models.productcarousel;
 
 import org.osgi.annotation.versioning.Version;
