+++ bundles/core/src/test/resources/graphql/magento-graphql-product-result.json
@@ -4,7 +4,6 @@
       "items": [
         {
           "__typename": "ConfigurableProduct",
-          "id": "1111",
           "sku": "MJ01",
           "name": "Beaumont Summit Kit",
           "description": {
@@ -132,7 +131,6 @@
                 }
               ],
               "product": {
-                "id": "1112",
                 "sku": "MJ01-XS-Orange",
                 "name": "Beaumont Summit Kit-XS-Orange",
                 "description": {
@@ -188,7 +186,6 @@
                 }
               ],
               "product": {
-                "id": "1113",
                 "sku": "MJ01-XS-Red",
                 "name": "Beaumont Summit Kit-XS-Red",
                 "description": {
@@ -244,7 +241,6 @@
                 }
               ],
               "product": {
-                "id": "1114",
                 "sku": "MJ01-XS-Yellow",
                 "name": "Beaumont Summit Kit-XS-Yellow",
                 "description": {
@@ -314,7 +310,6 @@
                 }
               ],
               "product": {
-                "id": "1115",
                 "sku": "MJ01-S-Orange",
                 "name": "Beaumont Summit Kit-S-Orange",
                 "description": {
@@ -370,7 +365,6 @@
                 }
               ],
               "product": {
-                "id": "1116",
                 "sku": "MJ01-S-Red",
                 "name": "Beaumont Summit Kit-S-Red",
                 "description": {
@@ -426,7 +420,6 @@
                 }
               ],
               "product": {
-                "id": "1117",
                 "sku": "MJ01-S-Yellow",
                 "name": "Beaumont Summit Kit-S-Yellow",
                 "description": {
@@ -496,7 +489,6 @@
                 }
               ],
               "product": {
-                "id": "1118",
                 "sku": "MJ01-M-Orange",
                 "name": "Beaumont Summit Kit-M-Orange",
                 "description": {
@@ -552,7 +544,6 @@
                 }
               ],
               "product": {
-                "id": "1119",
                 "sku": "MJ01-M-Red",
                 "name": "Beaumont Summit Kit-M-Red",
                 "description": {
@@ -608,7 +599,6 @@
                 }
               ],
               "product": {
-                "id": "1120",
                 "sku": "MJ01-M-Yellow",
                 "name": "Beaumont Summit Kit-M-Yellow",
                 "description": {
@@ -678,7 +668,6 @@
                 }
               ],
               "product": {
-                "id": "1121",
                 "sku": "MJ01-L-Orange",
                 "name": "Beaumont Summit Kit-L-Orange",
                 "description": {
@@ -734,7 +723,6 @@
                 }
               ],
               "product": {
-                "id": "1122",
                 "sku": "MJ01-L-Red",
                 "name": "Beaumont Summit Kit-L-Red",
                 "description": {
@@ -790,7 +778,6 @@
                 }
               ],
               "product": {
-                "id": "1123",
                 "sku": "MJ01-L-Yellow",
                 "name": "Beaumont Summit Kit-L-Yellow",
                 "description": {
@@ -860,7 +847,6 @@
                 }
               ],
               "product": {
-                "id": "1124",
                 "sku": "MJ01-XL-Orange",
                 "name": "Beaumont Summit Kit-XL-Orange",
                 "description": {
@@ -916,7 +902,6 @@
                 }
               ],
               "product": {
-                "id": "1125",
                 "sku": "MJ01-XL-Red",
                 "name": "Beaumont Summit Kit-XL-Red",
                 "description": {
@@ -972,7 +957,6 @@
                 }
               ],
               "product": {
-                "id": "1126",
                 "sku": "MJ01-XL-Yellow",
                 "name": "Beaumont Summit Kit-XL-Yellow",
                 "description": {
