+++ ui.apps/src/main/content/jcr_root/apps/core/cif/components/commerce/product/v1/product/clientlib/js/addToCart.js
@@ -71,7 +71,7 @@ class AddToCart {
 
         // Update sku attribute in select element
         document.querySelector(AddToCart.selectors.quantity).setAttribute('data-product-sku', variant.sku);
-        document.querySelector(AddToCart.selectors.quantity).setAttribute('data-product-id', 'product-' + variant.id);
+        document.querySelector(AddToCart.selectors.quantity).setAttribute('data-product-id', variant.id);
 
         // Update internal state
         this._state.sku = variant.sku;
